move types in compile-common to avoid clobbering
